pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "requests>=2.28.1",
    "tqdm>=4.64.1",
    "python-dotenv>=0.21.0",
]
requires-python = ">=3.11"
license = {text = ""}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
